feat: Add list-view button event
리스트 뷰 왼쪽 오른쪽 버튼 클릭 시 기사 변경
